Worked on user flow so that user could click Get Started to be taken to main Market Index Page, implemented a link to Sign Up on Login page and a Login link on the Sign Up page created a link back to Marketing Page on the Main Page implemented in the Logo. And added animation sytling to Login and Get Started Button on the Marketing Page.
